feat: version domain API
