+++ Functions/RIPA.Functions.Domain/Functions/v1/Templates/DeleteTemplate.cs
@@ -7,33 +7,33 @@ using Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Attributes;
 using Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Enums;
 using Microsoft.Extensions.Logging;
 using Microsoft.OpenApi.Models;
-using RIPA.Functions.Domain.Functions.Statutes.Models;
+using RIPA.Functions.Domain.Functions.v1.Templates.Models;
 using RIPA.Functions.Security;
 using System;
 using System.Net;
 using System.Threading.Tasks;
 
-namespace RIPA.Functions.Domain.Functions.Statutes;
+namespace RIPA.Functions.Domain.Functions.v1.Templates;
 
-public class DeleteStatute
+public class DeleteTemplate
 {
     private readonly TableServiceClient _tableServiceClient;
     private readonly TableClient _tableClient;
 
-    public DeleteStatute(TableServiceClient tableServiceClient)
+    public DeleteTemplate(TableServiceClient tableServiceClient)
     {
         _tableServiceClient = tableServiceClient;
-        _tableClient = _tableServiceClient.GetTableClient("Statutes");
+        _tableClient = _tableServiceClient.GetTableClient("Templates");
     }
 
-    [FunctionName("DeleteStatute")]
-    [OpenApiOperation(operationId: "DeleteStatute", tags: new[] { "name" })]
+    [FunctionName("v1/DeleteTemplate")]
+    [OpenApiOperation(operationId: "v1/DeleteTemplate", tags: new[] { "name", "v1" })]
     [OpenApiSecurity("Bearer", SecuritySchemeType.OAuth2, Name = "Bearer Token", In = OpenApiSecurityLocationType.Header, Flows = typeof(RIPAAuthorizationFlow))]
     [OpenApiParameter(name: "Ocp-Apim-Subscription-Key", In = ParameterLocation.Header, Required = true, Type = typeof(string), Description = "Ocp-Apim-Subscription-Key")]
-    [OpenApiParameter(name: "Id", In = ParameterLocation.Path, Required = true, Type = typeof(string), Description = "The Statute Id/Name")]
-    [OpenApiResponseWithBody(statusCode: HttpStatusCode.OK, contentType: "application/json", bodyType: typeof(string), Description = "Statute deleted")]
-    [OpenApiResponseWithBody(statusCode: HttpStatusCode.BadRequest, contentType: "application/json", bodyType: typeof(string), Description = "Statute Id not found")]
-    public async Task<IActionResult> Run([HttpTrigger(AuthorizationLevel.Function, "delete", Route = "DeleteStatute/{Id}")] HttpRequest req, int Id, ILogger log)
+    [OpenApiParameter(name: "Id", In = ParameterLocation.Path, Required = true, Type = typeof(string), Description = "The Template Id/Name")]
+    [OpenApiResponseWithBody(statusCode: HttpStatusCode.OK, contentType: "application/json", bodyType: typeof(string), Description = "Template deleted")]
+    [OpenApiResponseWithBody(statusCode: HttpStatusCode.BadRequest, contentType: "application/json", bodyType: typeof(string), Description = "Template Id not found")]
+    public async Task<IActionResult> Run([HttpTrigger(AuthorizationLevel.Function, "delete", Route = "v1/DeleteTemplate/{Id}")] HttpRequest req, string Id, ILogger log)
     {
         try
         {
@@ -50,14 +50,15 @@ public class DeleteStatute
 
         try
         {
-            Statute statute = new Statute { PartitionKey = "CA", RowKey = Id.ToString(), OffenseCode = Id };
-            var response = await _tableClient.DeleteEntityAsync(statute.PartitionKey, statute.RowKey);
+            Template template = new Template { PartitionKey = "CA", RowKey = Id };
+
+            var response = await _tableClient.DeleteEntityAsync(template.PartitionKey, template.RowKey);
 
             return new OkObjectResult(response.ToString());
         }
         catch
         {
-            return new BadRequestObjectResult("Statute Offense Code not found");
+            return new BadRequestObjectResult("Template not found");
         }
     }
 }
