+++ Functions/RIPA.Functions.Domain/Functions/v1/Templates/GetTemplates.cs
@@ -7,33 +7,33 @@ using Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Attributes;
 using Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Enums;
 using Microsoft.Extensions.Logging;
 using Microsoft.OpenApi.Models;
-using RIPA.Functions.Domain.Functions.Templates.Models;
+using RIPA.Functions.Domain.Functions.v1.Templates.Models;
 using RIPA.Functions.Security;
 using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
 
-namespace RIPA.Functions.Domain.Functions.Templates;
+namespace RIPA.Functions.Domain.Functions.v1.Templates;
 
-public class GetTemplate
+public class GetTemplates
 {
     private readonly TableServiceClient _tableServiceClient;
     private readonly TableClient _tableClient;
 
-    public GetTemplate(TableServiceClient tableServiceClient)
+    public GetTemplates(TableServiceClient tableServiceClient)
     {
         _tableServiceClient = tableServiceClient;
         _tableClient = _tableServiceClient.GetTableClient("Templates");
     }
 
-    [FunctionName("GetTemplate")]
-    [OpenApiOperation(operationId: "GetTemplate", tags: new[] { "name" })]
+    [FunctionName("v1/GetTemplates")]
+    [OpenApiOperation(operationId: "v1/GetTemplates", tags: new[] { "name", "v1" })]
     [OpenApiSecurity("Bearer", SecuritySchemeType.OAuth2, Name = "Bearer Token", In = OpenApiSecurityLocationType.Header, Flows = typeof(RIPAAuthorizationFlow))]
     [OpenApiParameter(name: "Ocp-Apim-Subscription-Key", In = ParameterLocation.Header, Required = true, Type = typeof(string), Description = "Ocp-Apim-Subscription-Key")]
-    [OpenApiParameter(name: "Id", In = ParameterLocation.Path, Required = true, Type = typeof(string), Description = "The Template Id/Name")]
-    [OpenApiResponseWithBody(statusCode: HttpStatusCode.OK, contentType: "application/json", bodyType: typeof(Template), Description = "Template Object")]
-    [OpenApiResponseWithBody(statusCode: HttpStatusCode.BadRequest, contentType: "application/json", bodyType: typeof(string), Description = "Template Id not found")]
-    public async Task<IActionResult> Run([HttpTrigger(AuthorizationLevel.Function, "get", Route = "GetTemplate/{Id}")] HttpRequest req, string Id, ILogger log)
+    [OpenApiResponseWithBody(statusCode: HttpStatusCode.OK, contentType: "application/json", bodyType: typeof(string), Description = "List of Template")]
+    public async Task<IActionResult> Run([HttpTrigger(AuthorizationLevel.Function, "get", Route = "v1/GetTemplates")] HttpRequest req, ILogger log)
     {
         try
         {
@@ -48,15 +48,22 @@ public class GetTemplate
             return new UnauthorizedResult();
         }
 
-        try
-        {
-            Template result = await _tableClient.GetEntityAsync<Template>("CA", Id);
+        List<Template> response = new List<Template>();
 
-            return new OkObjectResult(result);
-        }
-        catch
+        var queryResult = _tableClient.Query<Template>();
+
+        foreach (var template in queryResult)
         {
-            return new BadRequestObjectResult("Template not found");
+            if (template?.DeactivationDate <= DateTime.Now)
+            {
+                continue;
+            }
+            response.Add(template);
         }
+
+        response.OrderBy(t => t.Id);
+
+        log.LogInformation($"GetTemplates returned {response.Count} templates");
+        return new OkObjectResult(response);
     }
 }
