+++ Functions/RIPA.Functions.Domain/Functions/v1/Templates/PutTemplate.cs
@@ -7,13 +7,13 @@ using Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Attributes;
 using Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Enums;
 using Microsoft.Extensions.Logging;
 using Microsoft.OpenApi.Models;
-using RIPA.Functions.Domain.Functions.Templates.Models;
+using RIPA.Functions.Domain.Functions.v1.Templates.Models;
 using RIPA.Functions.Security;
 using System;
 using System.Net;
 using System.Threading.Tasks;
 
-namespace RIPA.Functions.Domain.Functions.Templates;
+namespace RIPA.Functions.Domain.Functions.v1.Templates;
 
 public class PutTemplate
 {
@@ -26,15 +26,15 @@ public class PutTemplate
         _tableClient = _tableServiceClient.GetTableClient("Statutes");
     }
 
-    [FunctionName("PutTemplate")]
-    [OpenApiOperation(operationId: "PutTemplate", tags: new[] { "name" })]
+    [FunctionName("v1/PutTemplate")]
+    [OpenApiOperation(operationId: "v1/PutTemplate", tags: new[] { "name", "v1" })]
     [OpenApiSecurity("Bearer", SecuritySchemeType.OAuth2, Name = "Bearer Token", In = OpenApiSecurityLocationType.Header, Flows = typeof(RIPAAuthorizationFlow))]
     [OpenApiParameter(name: "Ocp-Apim-Subscription-Key", In = ParameterLocation.Header, Required = true, Type = typeof(string), Description = "Ocp-Apim-Subscription-Key")]
     [OpenApiParameter(name: "Id", In = ParameterLocation.Path, Required = true, Type = typeof(int), Description = "The Template Id/Name")]
     [OpenApiRequestBody(contentType: "application/Json", bodyType: typeof(Template), Deprecated = false, Description = "Template object", Required = true)]
     [OpenApiResponseWithBody(statusCode: HttpStatusCode.OK, contentType: "application/json", bodyType: typeof(string), Description = "Template Created")]
     [OpenApiResponseWithBody(statusCode: HttpStatusCode.BadRequest, contentType: "application/json", bodyType: typeof(string), Description = "Template failed on insert or replace")]
-    public async Task<IActionResult> Run([HttpTrigger(AuthorizationLevel.Function, "Put", Route = "PutTemplate/{Id}")] Template template, HttpRequest req, string Id, ILogger log)
+    public async Task<IActionResult> Run([HttpTrigger(AuthorizationLevel.Function, "Put", Route = "v1/PutTemplate/{Id}")] Template template, HttpRequest req, string Id, ILogger log)
     {
         try
         {
