+++ UI/src/store/index.js
@@ -640,7 +640,7 @@ export default new Vuex.Store({
     deleteBeat({ dispatch, state }, beat) {
       return axios
         .delete(
-          `${state.apiConfig.apiBaseUrl}domain/v${state.version}/DeleteBeat/${beat.id}`,
+          `${state.apiConfig.apiBaseUrl}domain/v1/DeleteBeat/${beat.id}`,
           {
             headers: {
               'Ocp-Apim-Subscription-Key': state.apiConfig.apiSubscription,
@@ -660,7 +660,7 @@ export default new Vuex.Store({
     editBeat({ dispatch, state }, beat) {
       return axios
         .put(
-          `${state.apiConfig.apiBaseUrl}domain/v${state.version}/PutBeat/${beat.id}`,
+          `${state.apiConfig.apiBaseUrl}domain/v1/PutBeat/${beat.id}`,
           beat,
           {
             headers: {
@@ -739,17 +739,13 @@ export default new Vuex.Store({
       const formData = new FormData()
       formData.append('file', domainFile)
       return axios
-        .post(
-          `${state.apiConfig.apiBaseUrl}domain/v${state.version}/PostUpload`,
-          formData,
-          {
-            headers: {
-              'Content-Type': 'multipart/form-data',
-              'Ocp-Apim-Subscription-Key': state.apiConfig.apiSubscription,
-              'Cache-Control': 'no-cache',
-            },
+        .post(`${state.apiConfig.apiBaseUrl}domain/v1/PostUpload`, formData, {
+          headers: {
+            'Content-Type': 'multipart/form-data',
+            'Ocp-Apim-Subscription-Key': state.apiConfig.apiSubscription,
+            'Cache-Control': 'no-cache',
           },
-        )
+        })
         .then(response => {
           dispatch('getAdminSchools')
           dispatch('getAdminCities')
@@ -973,7 +969,7 @@ export default new Vuex.Store({
 
     getAdminBeats({ commit, state }) {
       return axios
-        .get(`${state.apiConfig.apiBaseUrl}domain/v${state.version}/GetBeats`, {
+        .get(`${state.apiConfig.apiBaseUrl}domain/v1/GetBeats`, {
           headers: {
             'Ocp-Apim-Subscription-Key': state.apiConfig.apiSubscription,
             'Cache-Control': 'no-cache',
@@ -1011,15 +1007,12 @@ export default new Vuex.Store({
         })
       } else {
         return axios
-          .get(
-            `${state.apiConfig.apiBaseUrl}domain/v${state.version}/GetBeats`,
-            {
-              headers: {
-                'Ocp-Apim-Subscription-Key': state.apiConfig.apiSubscription,
-                'Cache-Control': 'no-cache',
-              },
+          .get(`${state.apiConfig.apiBaseUrl}domain/v1/GetBeats`, {
+            headers: {
+              'Ocp-Apim-Subscription-Key': state.apiConfig.apiSubscription,
+              'Cache-Control': 'no-cache',
             },
-          )
+          })
           .then(response => {
             const data = response.data
               .sort((x, y) => {
@@ -1053,14 +1046,11 @@ export default new Vuex.Store({
 
     getAdminCities({ commit, state }) {
       return axios
-        .get(
-          `${state.apiConfig.apiBaseUrl}domain/v${state.version}/GetCities`,
-          {
-            headers: {
-              'Ocp-Apim-Subscription-Key': state.apiConfig.apiSubscription,
-            },
+        .get(`${state.apiConfig.apiBaseUrl}domain/v1/GetCities`, {
+          headers: {
+            'Ocp-Apim-Subscription-Key': state.apiConfig.apiSubscription,
           },
-        )
+        })
         .then(response => {
           const data = response.data
             .sort((x, y) => {
@@ -1093,14 +1083,11 @@ export default new Vuex.Store({
         })
       } else {
         return axios
-          .get(
-            `${state.apiConfig.apiBaseUrl}domain/v${state.version}/GetCities`,
-            {
-              headers: {
-                'Ocp-Apim-Subscription-Key': state.apiConfig.apiSubscription,
-              },
+          .get(`${state.apiConfig.apiBaseUrl}domain/v1/GetCities`, {
+            headers: {
+              'Ocp-Apim-Subscription-Key': state.apiConfig.apiSubscription,
             },
-          )
+          })
           .then(response => {
             const data = response.data.sort((x, y) => {
               const cityA = x.name.toUpperCase()
@@ -1153,14 +1140,11 @@ export default new Vuex.Store({
 
     getAdminSchools({ commit, state }) {
       return axios
-        .get(
-          `${state.apiConfig.apiBaseUrl}domain/v${state.version}/GetSchools`,
-          {
-            headers: {
-              'Ocp-Apim-Subscription-Key': state.apiConfig.apiSubscription,
-            },
+        .get(`${state.apiConfig.apiBaseUrl}domain/v1/GetSchools`, {
+          headers: {
+            'Ocp-Apim-Subscription-Key': state.apiConfig.apiSubscription,
           },
-        )
+        })
         .then(response => {
           const data = response.data
             .sort((x, y) => {
@@ -1194,14 +1178,11 @@ export default new Vuex.Store({
         })
       } else {
         return axios
-          .get(
-            `${state.apiConfig.apiBaseUrl}domain/v${state.version}/GetSchools`,
-            {
-              headers: {
-                'Ocp-Apim-Subscription-Key': state.apiConfig.apiSubscription,
-              },
+          .get(`${state.apiConfig.apiBaseUrl}domain/v1/GetSchools`, {
+            headers: {
+              'Ocp-Apim-Subscription-Key': state.apiConfig.apiSubscription,
             },
-          )
+          })
           .then(response => {
             const data = response.data
               .filter(item => item.status === 'Active')
@@ -1237,14 +1218,11 @@ export default new Vuex.Store({
 
     getAdminStatutes({ commit, state }) {
       return axios
-        .get(
-          `${state.apiConfig.apiBaseUrl}domain/v${state.version}/GetStatutes`,
-          {
-            headers: {
-              'Ocp-Apim-Subscription-Key': state.apiConfig.apiSubscription,
-            },
+        .get(`${state.apiConfig.apiBaseUrl}domain/v1/GetStatutes`, {
+          headers: {
+            'Ocp-Apim-Subscription-Key': state.apiConfig.apiSubscription,
           },
-        )
+        })
         .then(response => {
           const data = response.data.map(item => {
             return {
@@ -1322,15 +1300,12 @@ export default new Vuex.Store({
         })
       } else {
         return axios
-          .get(
-            `${state.apiConfig.apiBaseUrl}domain/v${state.version}/GetTemplates`,
-            {
-              headers: {
-                'Ocp-Apim-Subscription-Key': state.apiConfig.apiSubscription,
-                'Cache-Control': 'no-cache',
-              },
+          .get(`${state.apiConfig.apiBaseUrl}domain/v1/GetTemplates`, {
+            headers: {
+              'Ocp-Apim-Subscription-Key': state.apiConfig.apiSubscription,
+              'Cache-Control': 'no-cache',
             },
-          )
+          })
           .then(response => {
             const data = response.data.map(item => {
               const displayName = item.displayName
